Extract Issue#case_form_json to decorator
As already done with same method for other models. This is more of a
view concern, it just was first added before we started using Draper.
